project/requirements.txt
prefect==0.15.9
requests
pandas
psycopg2-binary